requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
base58==2.0.1
bitarray==1.2.2
certifi==2018.11.29
chardet==3.0.4
cytoolz==0.11.0
distlib==0.3.0
Django==3.1.7
django-restframework==0.0.1
djangorestframework==3.12.2
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
filelock==3.0.12
gspread==3.1.0
hexbytes==0.2.1
httplib2==0.15.0
idna==2.8
importlib-metadata==1.6.1
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
lru-dict==1.1.6
multiaddr==0.0.9
netaddr==0.8.0
oauth2client==4.1.3
parsimonious==0.8.1
pipenv==2020.6.2
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycryptodome==3.9.9
pyrsistent==0.17.3
pytz==2021.1
requests==2.22.0
rlp==2.0.1
rsa==4.0
six==1.13.0
sqlparse==0.4.1
toolz==0.11.1
typing-extensions==3.7.4.3
urllib3==1.25.7
varint==1.0.2
virtualenv==20.0.23
virtualenv-clone==0.5.4
web3==5.13.1
websockets==8.1
zipp==3.1.0
